Fix the position of close tip button.
So that if the content of the tip popup overflows the popup,
scrolling the content doesn't also scroll the close button out of view.

Before this patch, the entire popup was scrollable, but this patch
only makes the inner content scrollable, so that the close button stays
in a fixed position.