Hide "Review final grade" link by default.
The link is only used during extended feedback in assessment mode.  It's hidden
on page load by JS in all cases, and only shown if the student goes back to a
question after using up all attempts, so we can just as well hide it in CSS.

This fixes MCKIN-3099.