Tests for the two issues when student double-clicks on the submit button.
First issue is when user selects the right answer and then double clicks submit
expected outcome is that green checkmark is displayed, real outcome is that
red crossmark is displayed, as if the answer was wrong.

Second issue is when student double-clicks on the submit button, expected
outcome is that only a single attempt is "used", real outcome is that
two attempts are used.